+++ src/Modals/CreateTaskModal.ts
@@ -2,6 +2,7 @@ import { App } from "obsidian";
 import CreateFileModal from "./CreateFileModal";
 import AgileProjectPlugin from "main";
 import "../../styles.css";
+import VaultParser from "src/VaultParser";
 
 /**
  * Modal for creating a new Task in the Agile Project Plugin.
@@ -28,7 +29,7 @@ class CreateTaskModal extends CreateFileModal {
         contentEl.createEl('h1', { text: 'Creating Agile Task' });
 
         let nameInput = this.SingleLineInput(contentEl, 'Task Name :', 'Enter task name');
-        let epicDropdown = this.SingleLineDropdown(contentEl, 'Epic Name :', this.Plugin.StructureManager.GetEpics());
+        let epicDropdown = this.SingleLineDropdown(contentEl, 'Epic Name :', new VaultParser(this.app, this.Plugin).GetEpics());
         let storyDropdown = this.SingleLineDropdown(contentEl, 'Story Name :', []);
         let priorityDropdown = this.SingleLineDropdown(contentEl, 'Task Priority :', ['Low', 'Medium', 'High']);
         let descInput = this.MultiLineInput(contentEl, 'Task Description', 'Enter task description');
@@ -60,7 +61,7 @@ class CreateTaskModal extends CreateFileModal {
      */
     private UpdateStoryDropdown(epicName: string, storyDropdown: HTMLSelectElement): void {
         storyDropdown.empty();
-        this.Plugin.StructureManager.GetStories(epicName).forEach(story => {
+        new VaultParser(this.app, this.Plugin).GetStories(epicName).forEach(story => {
             storyDropdown.createEl('option', { text: story, value: story });
         });
     }
